updated styling
