+++ pages/_app.js
@@ -5,7 +5,12 @@ export default function Handcrafted({ Component, pageProps }) {
   const getLayout = Component.getLayout || ((page) => page)
 
   return ( 
-    <Theme appearance="dark" accentColor="sky" grayColor="mauve">
+    <Theme
+      appearance="dark"
+      accentColor="sky" 
+      grayColor="mauve"
+      panelBackground="solid"
+    >
       {getLayout(<Component {...pageProps} />)}
     </Theme>
   )
